ground probes shall be close to signal probes
when analyzing a 50MHz SPI NOR flash, we found that connecting a ground probe to a probe point far away from the SPI chip (for convenience; the ground probe point was easily accessible), the amount of glitches was very high. only by connecting a ground probe to the ground pin of the chip in question the glitches could be eliminated.